Refactor: split up.
Almost all of the methods don't actually have to be methods. Regular functions would work just as well.

Split up the (which started to become) giant single, monolithic class and host functionality in individual modules.

We'll create an object hierarchy at a later time (which will wrap the free functions) to bring back the OOP aspects.