Introduce new OOP wrapper.
This makes it somewhat easier to work with the various bits of
functionality offered by dominh.

Main entry point is the connect(..) function, which returns a Controller
instance. Through that instance, registers, frames and variables can be
accessed. Convenience classes keep track of indices and group nrs to avoid
the user having to provide those every time.